Revert "Feature/tfgdam2 10 creacion screen profile"
