+++ assets/adaptive-icon.png
