+++ package.json
@@ -9,30 +9,13 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "@expo/vector-icons": "^13.0.0",
-    "@react-navigation/bottom-tabs": "^6.5.7",
-    "@react-navigation/drawer": "^6.6.2",
-    "@react-navigation/native": "^6.1.6",
-    "@react-navigation/native-stack": "^6.9.12",
-    "@react-navigation/stack": "^6.3.16",
-    "@types/react": "~18.0.27",
-    "deprecated-react-native-prop-types": "^4.0.0",
     "expo": "^48.0.6",
-    "expo-constants": "~14.2.1",
-    "expo-linking": "~4.0.1",
-    "expo-router": "^1.2.2",
     "expo-status-bar": "~1.4.4",
-    "memoize-one": "^6.0.0",
     "react": "18.2.0",
-    "react-native": "0.71.3",
-    "react-native-gesture-handler": "~2.9.0",
-    "react-native-safe-area-context": "4.5.0",
-    "react-native-screens": "~3.20.0",
-    "typescript": "^4.9.4"
+    "react-native": "0.71.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.20.0",
-    "react-native-dev-menu": "^4.1.1"
+    "@babel/core": "^7.20.0"
   },
   "private": true
 }
