--- src/screens/bottom_navigation_screens/ProfileMenuScreen.js
@@ -1,9 +1,0 @@
-import React from 'react';
-import ProfileMenuNavigation from "../../navigation/ProfileStackNavigation";
-
-
-export default function ProfileMenuScreen() {
-  return (
-      <ProfileMenuNavigation/>
-  );
-}
